chore(admin): folder structure change for editing tiles
