+++ next-tavla/src/Admin/scenarios/TileSettings/components/QuaySettings.tsx
@@ -3,8 +3,7 @@ import { GetQuayQuery } from 'graphql/index'
 import { useQuery } from 'graphql/utils'
 import { TQuayTile } from 'types/tile'
 import { fieldsNotNull } from 'utils/typeguards'
-import { SortableTileWrapper } from './SortableTileWrapper'
-import { TileSettingsWrapper } from './TileSettingsHeader'
+import { TileSettingsWrapper } from '../../TileSettings/components/TileSettingsWrapper'
 
 function QuaySettings({ tile }: { tile: TQuayTile }) {
     const { data } = useQuery(GetQuayQuery, { quayId: tile.placeId })
@@ -18,11 +17,9 @@ function QuaySettings({ tile }: { tile: TQuayTile }) {
           (data.quay?.description ?? data.quay?.publicCode)
 
     return (
-        <SortableTileWrapper id={tile.uuid}>
-            <TileSettingsWrapper uuid={tile.uuid} name={name}>
-                <SelectLines tile={tile} lines={lines} />
-            </TileSettingsWrapper>
-        </SortableTileWrapper>
+        <TileSettingsWrapper uuid={tile.uuid} name={name}>
+            <SelectLines tile={tile} lines={lines} />
+        </TileSettingsWrapper>
     )
 }
 export { QuaySettings }
