+++ next-tavla/src/Admin/scenarios/TileSettings/components/StopPlaceSettings.tsx
@@ -3,7 +3,7 @@ import { StopPlaceSettingsQuery } from 'graphql/index'
 import { useQuery } from 'graphql/utils'
 import { TStopPlaceTile } from 'types/tile'
 import { fieldsNotNull } from 'utils/typeguards'
-import { TileSettingsWrapper } from './TileSettingsHeader'
+import { TileSettingsWrapper } from './TileSettingsWrapper'
 
 function StopPlaceSettings({ tile }: { tile: TStopPlaceTile }) {
     const { data } = useQuery(StopPlaceSettingsQuery, { id: tile.placeId })
