+++ next-tavla/src/Admin/scenarios/TileSettings/components/TileSettingsWrapper/styles.module.css
@@ -1,0 +1,15 @@
+.heading {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding-bottom: 1em;
+}
+
+.overviewWrapper {
+    background-color: var(--table-border-color);
+    border-radius: 0.5em;
+    padding: 1em;
+    width: 50em;
+    height: 20em;
+    overflow: scroll;
+}
