+++ next-tavla/src/Admin/scenarios/TilesOverview/components/DeleteButton/index.tsx
