--- next-tavla/src/Admin/scenarios/TilesSettings/components/TileSettingsHeader/styles.module.css
@@ -1,6 +1,0 @@
-.heading {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding-bottom: 1em;
-}
